Improve Python patterns and update to new SDK
