--- python/patterns-use-cases/hypercorn-config.toml
@@ -1,6 +1,0 @@
-bind = "0.0.0.0:9080"
-h2_max_concurrent_streams = 2147483647
-keep_alive_max_requests = 2147483647
-keep_alive_timeout = 2147483647
-workers = 8
-
