+++ python/patterns-use-cases/requirements.txt
@@ -1,5 +1,5 @@
 hypercorn
 restate_sdk==0.6.0
 pydantic
-requests
+httpx
 stripe
