--- python/patterns-use-cases/src/statemachinepayments/data_types.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class Result(BaseModel):
-    success: bool
-    message: str
