+++ python/patterns-use-cases/statemachinepayments/app.py
@@ -1,0 +1,12 @@
+import restate
+from accounts import account
+from payment_processor import payment_processor
+
+app = restate.app([payment_processor, account])
+
+if __name__ == "__main__":
+    import hypercorn
+    import asyncio
+    conf = hypercorn.Config()
+    conf.bind = ["0.0.0.0:9080"]
+    asyncio.run(hypercorn.asyncio.serve(app, conf))
